==> Code Library/2-Merging Data.ipynb <==
# Creating Master Dataframe for Modeling

import pandas as pd
import re

### Load Data

# Load each CSV file
patient_df = pd.read_csv('./Data Library/data/patient.csv')
dexcom_cgm_df = pd.read_csv('./Data Library/data/dexcom_cgm.csv')

# Inspect each DataFrame to determine the best way to combine them
print("Patient Data:", patient_df.shape)
print("Dexcom CGM Data:", dexcom_cgm_df.shape)

## Merging Patient Data

# Patient-Based Keys have standardized formatting
patient_df['Age'] = patient_df['Age'].astype(int)  # integer format
patient_df['Sex'] = patient_df['Sex'].astype(int)  # Standardize 'Sex' column

patient_df.columns
patient_df.shape

dexcom_cgm_df.columns
dexcom_cgm_df.shape

# Calculate the necessary repetitions of patient data to join with glucose data
num_repeats = len(dexcom_cgm_df) // len(patient_df) + 1
expanded_patient_df = pd.concat([patient_df] * num_repeats, ignore_index=True)
expanded_patient_df = expanded_patient_df[:len(dexcom_cgm_df)]  # Truncate to match glucose data length

# Combine patient data with Dexcom glucose data (should result in 17118 patient records)
simulated_df = pd.concat([expanded_patient_df.reset_index(drop=True), dexcom_cgm_df.reset_index(drop=True)], axis=1)

simulated_df['Time Checked'] = simulated_df['displayTime']
patient = simulated_df.drop(columns=['displayTime'])

# Display the simulated dataframe
patient.head()

# Validate record count
patient.shape

# Export the final aggregated data to a CSV file
output_path = './Data Library/data/patient_records.csv'
patient.to_csv(output_path, index=False)

print(f"Final combined data exported to: {output_path}")
