==> Code Library/4a-Preparation for Modeling.ipynb <==
# Scoring and Preparation for Modeling

### Imports

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

#### Read in Preprocessed Data

# Restaurant Menu Data
menu_df = pd.read_csv('./Data Library/preprocessed_data/menu_df.csv')
menu_df.head()

# Individual Food Data
individual_foods_df = pd.read_csv('./Data Library/preprocessed_data/individual_foods_df.csv')

individual_foods_df.head()

# Patient Data
patient_df = pd.read_csv('./Data Library/preprocessed_data/patient_df.csv')
patient_df.head()

### Compute Scores for Recommendations

While we already have a score that captures the suitability of a meal/food item based on general diabetes nutrition recommendations, this patient-specific score uses that general score in combination with personalized factors to suggest the best meal options for a person.

# # Patient-Specific Score Function
def patient_suitability_score(patient_data, food_data):
    # Patient Info
    glucose_rank = patient_data['GlucoseRank']
    high_bp = patient_data['HighBP']
    high_chol = patient_data['HighChol']
    smoker = patient_data['Smoker']
    stroke = patient_data['Stroke']
    heart = patient_data['HeartDiseaseorAttack']
    bmi = patient_data['BMI']

    # Food Info
    carbohydrates = food_data['carbohydrates']
    sugars = food_data['sugars']
    fiber = food_data['fiber']
    proteins = food_data['proteins']
    # None for individual food, where sodium isn't available
    sodium = food_data.get('sodium', None)
    fat = food_data['fats']

    # Base weights
    base_weight = np.array([0.45, 0.2, 0.15, 0.1, 0.05, 0.05])
    # Base nutrition goals (carbs, sugar, fiber, protein, sodium, fat)
    goals = np.array([52.5, 7.5, 7.5, 20, 765, 25])

    # Glucose Level Adjustments
    if glucose_rank == 'High':
        goals[1] = 2.5 
    elif glucose_rank == 'Low':
        goals[1] = 1.5

    # Adjustments for high blood pressure
    if high_bp:
        goals[1] = 2.5  # Adjust sugar goal
        goals[4] = 500  # Adjust sodium goal
        base_weight += np.array([0.3, 0.25, 0.15, 0.05, 0.2, 0.05])

    # Adjustments for high cholesterol
    if high_chol:
        goals[1] = 2.5  # Adjust sugar goal
        goals[2] = 10   # Adjust fiber goal
        goals[5] = 15   # Adjust fat goal
        base_weight += np.array([0.3, 0.25, 0.25, 0.1, 0.05, 0.05])

    # Adjustments for smokers
    if smoker:
        base_weight += np.array([0.25, 0.25, 0.25, 0.1, 0.1, 0.05])

    # Adjustments for stroke or heart attacks
    if stroke or heart:
        base_weight += np.array([0.25, 0.15, 0.25, 0.1, 0.15, 0.1])

    # Average Weights
    avg_weights = base_weight / base_weight.sum()

    # Scores Calculation
    scores = np.array([
        1 - abs(carbohydrates - goals[0]) / goals[0],  
        1 - sugars / goals[1],                        
        min(fiber / goals[2], 1),                     
        min(proteins / goals[3], 1),                  
        1 - (sodium / goals[4]) if sodium is not None else 1,  
        1 - fat / goals[5]                            
    ])

    scores = np.maximum(scores, 0)

    # Weighted Score
    final_score = (avg_weights * scores).sum()

    # BMI Penalty
    bmi_penalty = 0.025 * (bmi - 25)
    final_score = max(0, final_score - bmi_penalty)

    return round(final_score, 3)

#### Menu Data Patient Scores

# Initialize empty df for recommendations
menu_recs = []

# Run each patient/menu food pair through the function to calculate a score
for _, patient in patient_df.iterrows():
    for _, food in menu_df.iterrows():
        score = patient_suitability_score(patient, food)

        menu_recs.append({
            # Patient Columns
            'Patient_ID': patient['ID'],
            'HighBP': patient['HighBP'],
            'HighChol': patient['HighChol'],
            'CholCheck': patient['CholCheck'],
            'BMI': patient['BMI'],
            'Smoker': patient['Smoker'],
            'Stroke': patient['Stroke'],
            'HeartDiseaseorAttack': patient['HeartDiseaseorAttack'],
            'PhysActivity': patient['PhysActivity'],
            'GenHlth': patient['GenHlth'],
            'MentHlth': patient['MentHlth'],
            'PhysHlth': patient['PhysHlth'],
            'Sex': patient['Sex'],
            'Age': patient['Age'],
            'Glucose Value': patient['Glucose Value'],
            'Time Checked': patient['Time Checked'],
            'GlucoseRank': patient['GlucoseRank'],
            'Cluster': patient['Cluster'],

            # Food Columns
            'Restaurant': food['restaurant_name'],
            'Food_Name': food['food_name'],
            'Calories': food['calories'],
            'Carbohydrates': food['carbohydrates'],
            'Sugars': food['sugars'],
            'Fats': food['fats'],
            'Saturated_Fats': food['saturated_fats'],
            'Cholesterol': food['cholesterol'],
            'Sodium': food['sodium'],
            'Fiber': food['fiber'],
            'Potassium': food['potassium'],
            'Proteins': food['proteins'],
            'General_Score': food['score'],
            'Patient_Score': score
        })

# Convert to pandas DataFrame
menu_recs_df = pd.DataFrame(menu_recs)

# Display first 5 records
menu_recs_df.head()

#### Individual Food Patient Scores

# Initialize empty df for recommendations
individual_food_recs = []

# Run each patient/individual food pair through the function to calculate a score
for _, patient in patient_df.iterrows():
    for _, food in individual_foods_df.iterrows():
        score = patient_suitability_score(patient, food)

        individual_food_recs.append({
            # Patient Columns
            'Patient_ID': patient['ID'],
            'HighBP': patient['HighBP'],
            'HighChol': patient['HighChol'],
            'CholCheck': patient['CholCheck'],
            'BMI': patient['BMI'],
            'Smoker': patient['Smoker'],
            'Stroke': patient['Stroke'],
            'HeartDiseaseorAttack': patient['HeartDiseaseorAttack'],
            'PhysActivity': patient['PhysActivity'],
            'GenHlth': patient['GenHlth'],
            'MentHlth': patient['MentHlth'],
            'PhysHlth': patient['PhysHlth'],
            'Sex': patient['Sex'],
            'Age': patient['Age'],
            'Glucose Value': patient['Glucose Value'],
            'Time Checked': patient['Time Checked'],
            'GlucoseRank': patient['GlucoseRank'],
            'Cluster': patient['Cluster'],

            # Food Columns
            'Food_Name': food['food_name'],
            'Calories': food['calories'],
            'Carbohydrates': food['carbohydrates'],
            'Sugars': food['sugars'],
            'Fats': food['fats'],
            'Fiber': food['fiber'],
            'Proteins': food['proteins'],
            'General_Score': food['score'],
            'Patient_Score': score
        })


# Convert to pandas DataFrame
individual_food_recs_df = pd.DataFrame(individual_food_recs)

# Display first 5 records
individual_food_recs_df.head()

#### Recommend Top 5 Meals and Top 5 Foods per Patient

# Group by Patient_ID and get the top 5 Patient_Score for menu recommendations
top_menu_recs = menu_recs_df.groupby('Patient_ID').apply(
    lambda group: group.nlargest(5, 'Patient_Score')
).reset_index(drop=True)

# Display the result
top_menu_recs.head()

# Group by Patient_ID and get the top 5 Patient_Score for menu recommendations
top_foods_recs = individual_food_recs_df.groupby('Patient_ID').apply(
    lambda group: group.nlargest(5, 'Patient_Score')
).reset_index(drop=True)

# Display the result
top_foods_recs.head()

# Stratify Menu Recommendation Data

# Perform stratified sampling
_, menu_recs_samp = train_test_split(
    menu_recs_df,
    test_size=0.01,
    stratify=menu_recs_df['Patient_ID'],
    random_state=42)

# Print sample head
print(menu_recs_samp.shape)
menu_recs_samp.head()

# Stratify Individual Food Recommendation Data

# Perform stratified sampling
_, individual_foods_samp = train_test_split(
    individual_food_recs_df,
    test_size=0.01,
    stratify=individual_food_recs_df['Patient_ID'],
    random_state=42)

# Print sample head
print(individual_foods_samp.shape)
individual_foods_samp.head()

# Save the Recommendations DataFrames as a CSV file in repo
data_dir = './Data Library/preprocessed_data/'
menu_recs_samp.to_csv(data_dir + 'menu_recs_samp.csv', index=False)
individual_foods_samp.to_csv(data_dir + 'individual_foods_samp.csv', index=False)
